physio_shop/requirements.txt
# FastAPI & ASGI server
fastapi[standard]>=0.115.0
uvicorn[standard]>=0.30.0

# Database
sqlalchemy[asyncio]>=2.0.30
aiosqlite>=0.20.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.9

# Migrations
alembic>=1.13.0

# Auth & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9

# Settings
pydantic-settings>=2.2.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
